refactor: query 관련 api controller service 분리 작업
