+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/spec/controller/SpecController.java
@@ -1,14 +1,16 @@
 package kakaotech.bootcamp.respec.specranking.domain.spec.controller;
 
 import kakaotech.bootcamp.respec.specranking.domain.spec.dto.request.PostSpecRequest;
-import kakaotech.bootcamp.respec.specranking.domain.spec.dto.response.PostSpecResponse;
 import kakaotech.bootcamp.respec.specranking.domain.spec.service.SpecPostPutService;
+import kakaotech.bootcamp.respec.specranking.global.dto.SimpleResponseDto;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestPart;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -20,21 +22,23 @@ public class SpecController {
     private final SpecPostPutService specService;
 
     @PostMapping
-    public PostSpecResponse createSpec(
+    @ResponseStatus(HttpStatus.CREATED)
+    public SimpleResponseDto createSpec(
             @RequestPart("spec") PostSpecRequest request,
             @RequestPart(value = "portfolioFile", required = false) MultipartFile portfolioFile) {
 
         specService.createSpec(request, portfolioFile);
-        return new PostSpecResponse(true, "스펙 입력 성공!");
+        return new SimpleResponseDto(true, "스펙 입력 성공");
     }
 
     @PutMapping("/{specId}")
-    public PostSpecResponse updateSpec(
+    public SimpleResponseDto updateSpec(
             @PathVariable Long specId,
             @RequestPart("spec") PostSpecRequest request,
             @RequestPart(value = "portfolioFile", required = false) MultipartFile portfolioFile) {
 
         specService.updateSpec(specId, request, portfolioFile);
-        return new PostSpecResponse(true, "스펙 수정 성공!");
+        return new SimpleResponseDto(true, "스펙 수정 성공");
     }
+
 }
