+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/spec/service/SpecQueryService.java
@@ -8,6 +8,7 @@ import java.util.Set;
 import kakaotech.bootcamp.respec.specranking.domain.bookmark.repository.BookmarkRepository;
 import kakaotech.bootcamp.respec.specranking.domain.comment.repository.CommentRepository;
 import kakaotech.bootcamp.respec.specranking.domain.spec.dto.response.RankingResponse;
+import kakaotech.bootcamp.respec.specranking.domain.spec.dto.response.SearchResponse;
 import kakaotech.bootcamp.respec.specranking.domain.spec.entity.Spec;
 import kakaotech.bootcamp.respec.specranking.domain.spec.repository.SpecRepository;
 import kakaotech.bootcamp.respec.specranking.domain.user.entity.User;
@@ -19,7 +20,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
-public class SpecGetService {
+public class SpecQueryService {
 
     private final SpecRepository specRepository;
     private final BookmarkRepository bookmarkRepository;
@@ -85,6 +86,69 @@ public class SpecGetService {
         }
     }
 
+    private String encodeCursor(Long id) {
+        return Base64.getEncoder().encodeToString(String.valueOf(id).getBytes());
+    }
+
+    public SearchResponse searchByNickname(String keyword, String cursor, int limit) {
+        try {
+            Long currentUserId = UserUtils.getCurrentUserId();
+
+            Long cursorId = decodeCursor(cursor);
+
+            List<Spec> specs = specRepository.searchByNickname(keyword, cursorId, limit + 1);
+
+            boolean hasNext = specs.size() > limit;
+            if (hasNext) {
+                specs = specs.subList(0, limit);
+            }
+
+            String nextCursor = hasNext && !specs.isEmpty() ? encodeCursor(specs.get(specs.size() - 1).getId()) : null;
+
+            Set<Long> bookmarkedSpecIds = bookmarkRepository.findSpecIdsByUserId(currentUserId);
+
+            Map<String, Integer> jobFieldUserCountMap = specRepository.countByJobFields();
+
+            List<SearchResponse.SearchResult> searchResults = new ArrayList<>();
+            int overallRank = 1;
+
+            for (Spec spec : specs) {
+                User user = spec.getUser();
+                String specJobField = spec.getWorkPosition();
+
+                int jobFieldRank = specRepository.findRankByJobField(spec.getId(), specJobField);
+
+                double averageScore = calculateAverageScore(spec);
+
+                int commentsCount = commentRepository.countBySpecId(spec.getId());
+
+                int bookmarksCount = bookmarkRepository.countBySpecId(spec.getId());
+
+                SearchResponse.SearchResult item = new SearchResponse.SearchResult();
+                item.setUserId(user.getId());
+                item.setNickname(user.getNickname());
+                item.setProfileImageUrl(user.getUserProfileUrl());
+                item.setSpecId(spec.getId());
+                item.setJobField(specJobField);
+                item.setAverageScore(averageScore);
+                item.setRankByJobField(jobFieldRank);
+                item.setTotalUsersCountByJobField(jobFieldUserCountMap.getOrDefault(specJobField, 0));
+                item.setRank(overallRank++);
+                item.setBookmarked(bookmarkedSpecIds.contains(spec.getId()));
+                item.setCommentsCount(commentsCount);
+                item.setBookmarksCount(bookmarksCount);
+
+                searchResults.add(item);
+            }
+
+            return SearchResponse.success(keyword, searchResults, hasNext, nextCursor);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return SearchResponse.fail("검색 중 오류가 발생했습니다.");
+        }
+    }
+
+
     private double calculateAverageScore(Spec spec) {
         return (spec.getEducationScore()
                 + spec.getWorkExperienceScore()
@@ -105,9 +169,5 @@ public class SpecGetService {
         } catch (Exception e) {
             return Long.MAX_VALUE;
         }
-    }
-
-    private String encodeCursor(Long id) {
-        return Base64.getEncoder().encodeToString(String.valueOf(id).getBytes());
     }
 }
