+++ src/main/java/kakaotech/bootcamp/respec/specranking/global/dto/SimpleResponseDto.java
@@ -1,0 +1,15 @@
+package kakaotech.bootcamp.respec.specranking.global.dto;
+
+import lombok.Data;
+
+@Data
+public class SimpleResponseDto {
+    private boolean isSuccess;
+    private String message;
+
+    public SimpleResponseDto(boolean isSuccess, String message) {
+        this.isSuccess = isSuccess;
+        this.message = message;
+    }
+}
+
